Uninitialised width in screen geometry - found by iss
Another bug in the initialisation which would result in memory being trampled (albeit ROM).  The fix is to modify the initialisation table gr_scrngeom_base so that the base and text values are initialised together.
Also there was found unused gr_text_size word - removed from code and include file updated to be temp (struct is still the same size so any pokes will still work).